refactor(csharp): move CallAsyncSSEApi into Client.cs
Keep the hand-written async SSE entry next to CallSSEApi and drop the extra partial file per review feedback.
