+++ csharp/ChangeLog.md
@@ -1,6 +1,6 @@
 ### 2026-07-20 Version 0.2.2
 
-- Support CallAsyncSSEApi on netstandard2.1+ (IAsyncEnumerable; not generated by darabonba).
+- Support CallAsyncSSEApi on netstandard2.1+ (IAsyncEnumerable; hand-written in Client.cs, not generated by darabonba).
 - Add netstandard2.1 target framework.
 
 ### 2025-12-30 Version 0.2.1
